refactor: move flags to a separated file
